refactor(tool-call-repair): extract JSON tool-call parsing to json-tool-call.ts
Move scanJsonObject, parseJsonToolCallBlocksAt, tryClassifyJsonBuffer,
shouldStripJsonBlocks, looksLikeJsonToolCall, eventTemplate, and
createSyntheticTextDelta to a dedicated module to stay within LOC
limits for payload.ts, stream-normalizer.ts, and handlers.messages.ts.

Also extract asRecord, extractStandaloneMessageToolText, and
isMessageToolEnvelope to message-tool-helpers.ts in src/agents/.

New files are under 500 lines; oversized legacy files no longer grow.
