+++ packages/tool-call-repair/src/stream-normalizer.ts
@@ -12,8 +12,11 @@ import {
   utf8ByteLengthWithinLimit,
 } from "./grammar.js";
 import {
-  parseJsonToolCallBlocksAt,
-  scanJsonObject,
+  createSyntheticTextDelta,
+  eventTemplate,
+  tryClassifyJsonBuffer,
+} from "./json-tool-call.js";
+import {
   scanPlainTextToolCall,
   type PlainTextToolCallNameMatcher,
   type PlainTextToolCallScan,
@@ -424,22 +427,6 @@ export function projectScrubbedPlainTextToolCallMessage(params: {
     : undefined;
 }
 
-// Fast heuristic: does partial JSON text resemble a tool-call object?
-// When stream chunks start with `{` at line start we must decide whether to
-// buffer (withhold from visible text) or emit immediately.  Buffering a
-// false positive is cheap — the classifyPending gate replays non-tool-call
-// JSON once the object completes.
-function looksLikeJsonToolCall(text: string): boolean {
-  return (
-    text.includes('"name"') ||
-    text.includes('"tool_calls"') ||
-    text.includes('"toolCalls"') ||
-    text.includes('"function"') ||
-    text.includes('"tool_name"') ||
-    text.includes('"function_name"')
-  );
-}
-
 function findPotentialCallStart(
   text: string,
   atLineStart: boolean,
@@ -453,15 +440,9 @@ function findPotentialCallStart(
       continue;
     }
     const start = skipLineIndentation(text, index);
-    // Bare JSON tool-call objects: bracket/XML/Harmony scanners don't
-    // recognize a line-start `{`, so we must capture them here.  Accept
-    // any line-start `{` even when the shape key is not yet visible in
-    // the current fragment (e.g. first chunk is `{"na`).  classifyPending
-    // keeps buffering until the JSON completes, then validates the shape
-    // or replays as a false positive — at most one extra classify cycle.
     if (text[start] === "{") {
       return index;
-    }
+    } // Bare JSON: bracket/XML/Harmony don't match `{`
     const scan = scanPlainTextToolCall(text, start, {
       matcher,
       maxPayloadBytes: MAX_PAYLOAD_BYTES,
@@ -479,28 +460,6 @@ function nextAtLineStart(previous: boolean, text: string): boolean {
     return previous;
   }
   return text.endsWith("\n") || text.endsWith("\r");
-}
-
-function eventTemplate(event: Record<string, unknown>): Record<string, unknown> {
-  const template = { ...event };
-  delete template.content;
-  delete template.delta;
-  delete template.partial;
-  return template;
-}
-
-function createSyntheticTextDelta(
-  template: Record<string, unknown>,
-  text: string,
-  partial?: Record<string, unknown>,
-): Record<string, unknown> {
-  const event = eventTemplate(template);
-  return {
-    ...event,
-    type: "text_delta",
-    delta: text,
-    ...(partial ? { partial } : {}),
-  };
 }
 
 function cappedUtf8ByteLength(text: string): number {
@@ -787,57 +746,10 @@ function classifyPending(
 ): PendingClassification {
   const candidate = { text: pending.buffer, parts: pending.parts };
   const view = createCandidateScanView(candidate);
-
-  // Pre-check: bare JSON tool-call objects.  The bracket/XML/Harmony
-  // scanner does not recognize line-start `{`, so we must validate and
-  // classify buffered JSON candidates here.  When the JSON completes
-  // as a valid tool call the normalizer keeps it for terminal promotion;
-  // a non-tool-call shape is replayed as a false-positive so ordinary
-  // JSON prose is never dropped.
-  //
-  // Shape-key detection is deferred until the JSON object completes so
-  // that stream chunks split before or inside a quoted key (e.g. `{"na`)
-  // are buffered rather than leaked as visible text.  Incomplete JSON is
-  // held until the next chunk arrives or the buffer exceeds the size cap.
-  const jsonStart = skipLineIndentation(view.text, 0);
-  if (view.text[jsonStart] === "{") {
-    const json = scanJsonObject(view.text, jsonStart);
-    if (json.kind === "prefix") {
-      return pending.bufferBytes > MAX_PAYLOAD_BYTES
-        ? { kind: "false-positive" }
-        : { kind: "incomplete" };
-    }
-    // Complete JSON object — validate shape before trying to parse blocks.
-    // looksLikeJsonToolCall is the cheap pre-filter; parseJsonToolCallBlocksAt
-    // does the full structural validation including argument presence.
-    if (looksLikeJsonToolCall(view.text.slice(jsonStart))) {
-      const blocks = parseJsonToolCallBlocksAt(view.text, jsonStart);
-      if (blocks !== null && blocks.length > 0) {
-        const afterJson = skipWhitespace(view.text, json.end);
-        if (afterJson < view.text.length) {
-          // Trailing text after a complete JSON tool call.  The normalizer
-          // may duplicate the buffer when a suppress→candidate transition
-          // re-appends the same chunk; treat the stronger leading match as
-          // authoritative instead of yielding the trailing copy as text.
-          const trailingJsonStart = skipLineIndentation(view.text, afterJson);
-          if (
-            view.text[trailingJsonStart] === "{" &&
-            parseJsonToolCallBlocksAt(view.text, trailingJsonStart) !== null
-          ) {
-            return { kind: "complete" };
-          }
-          return { kind: "stripped", text: view.text.slice(afterJson) };
-        }
-        if (jsonStart > 0) {
-          return { kind: "stripped", text: view.text.slice(0, jsonStart) };
-        }
-        return { kind: "complete" };
-      }
-    }
-    // Complete JSON that is not a recognized tool call — replay as text.
-    return { kind: "false-positive" };
+  const jsonClass = tryClassifyJsonBuffer(view.text, pending.bufferBytes, MAX_PAYLOAD_BYTES); // Bare JSON
+  if (jsonClass) {
+    return jsonClass;
   }
-
   const terminalScan = scanPlainTextToolCall(view.text, skipLineIndentation(view.text, 0), {
     matcher,
     maxPayloadBytes: MAX_PAYLOAD_BYTES,
